REFACT: Clean up the code and fix issues
