+++ src/policy_checker.py
@@ -32,16 +32,17 @@
 
 # TODO: Check code reviews against review policies
 
+import os
 import argparse
 import re
 from sys import exit
 
-from code_review_policy import *
 from commit_manager import *
 from constants import *
-from crp_manager import *
 from github_API import *
 from gerrit_API import *
+from github_crp_manager import *
+from gerrit_crp_manager import *
 from review_manager import *
 
 
@@ -50,66 +51,25 @@ def github_validate_reviews(merge_commits, review_units):
     return True
 
 
-def check_gerrit_labels(crp, review_units):
-    #TODO check if rule is met
-    # https://github.com/GerritCodeReview/gerrit/blob/master/java/com/google/gerrit/server/rules/DefaultSubmitRule.java#L107
-    # https://github.com/GerritCodeReview/gerrit/blob/master/java/com/google/gerrit/entities/LabelFunction.java#L91-#L123
-
-    #extract the default policy
-    label = gerrit_extract_labels(crp)
-    rule = GERRIT_LABELS[label.upper()]
-
-    project_config = parse_crp(crp)[CONFIG_PROJECT]
-
-    status = "MAY"
-    if rule["isRequired"]:
-        status = "NEED"
-
-    for item in review_units:
-        signature, review = split_review_unit(item)
-        comment, score, reviewer = parse_review(review)
-
-        if score == '0':
-            continue
-
-        if rule["isBlock"] and is_max_negative(project_config, score):
-            if not is_allowed_to_block(crp, reviewer):
-                exit("Committer is not allowed to block the change")
-            status = "REJECT"
-            return status
-
-        if is_max_positive(project_config, score) or not rule["requiresMaxValue"]:
-            if not is_allowed_to_approve(crp, reviewer):
-                exit("Committer is not allowed to approve the change")
-            status = "MAY"
-
-            if rule["isRequired"]:
-                status = "OK"
-
-        #check for status
-        #https://github.com/GerritCodeReview/gerrit/blob/master/java/com/google/gerrit/server/rules/DefaultSubmitRule.java#L110
-
-    return status
-
-
 # Check if the reviews created on GitHub are legitimate
 def gerrit_validate_reviews(crp, review_units):
+
+    # Check if there are customized rules
+    rules_pl = gerrit_parse_crp(crp)[CONFIG_RULES]
+    if rules_pl != '':
+        # FIXME:
+        exit('Customized rules!')
+
     # Check if the basic rules (based on labels) are met
-    if not check_gerrit_labels(crp, review_units):
+    # Make decision based on the status
+    #https://github.com/GerritCodeReview/gerrit/blob/master/java/com/google/gerrit/server/rules/DefaultSubmitRule.java#L110
+    status = is_submittable(crp, review_units)
+    # TODO: Ensure it works for any situations
+    if status != "OK":
         return False
 
-    # Check if the minimum number of approving reviews is met
-    if not check_min_approvals(crp, review_units):
-        return False
-
-    # Check if there are required reviews from specific users
-    if not check_required_reviews(crp, review_units):
-        return False
-
-    # Check if stale approving reviews are dismissed
-    if not check_review_dismissal (crp, review_units):
-        return False
-
+    # TODO: Check for other policies if needed
+    
     return True
 
 
@@ -146,11 +106,16 @@ def validate_reviews(server, crp, merge_commit_type, merge_commits, review_units
 
 # Validate all reviews in a branch
 def validate_branch(server, repo, branch):
+    # Extract the repo name from the repo path
+    # e.g., 'd1/d2/f1' returns 'f1'
+    repo_name = os.path.basename(repo)
+
+    # Validate the CRP signature
     valid = True
     if server == GITHUB:
-        crp, valid = validate_github_crp(repo, branch)
+        crp, valid = validate_github_crp(repo_name, branch)
     elif server == GERRIT:
-        crp, valid = validate_gerrit_crp(repo, branch)
+        crp, valid = validate_gerrit_crp(repo_name, branch)
     else:
         exit(f"{server} is not supported!")
 
@@ -207,7 +172,7 @@ def create_parser():
         help='the path to the repository', )
         
     parser.add_argument('-b', '--branch', type=str, required=True,
-    help='the repository branch', )
+        help='the repository branch', )
 
     parser.add_argument('-s', '--server', type=str, required=True,
         help='the code review server either GitHub or Gerrit', )
@@ -228,12 +193,13 @@ def main():
     # Parse arguments
     parser = create_parser()
     args = parser.parse_args()
+
     server = args.server.lower()
-    repo = args.repo
     branch = args.branch
+    repo_path = args.repo
     
     # Validate the branch
-    validate_branch(server, repo, branch)
+    validate_branch(server, repo_path, branch)
 
 
 if __name__ == "__main__":
